+++ src/scripts/train/OLMo3-7B-noqknorm.py
@@ -1,6 +1,7 @@
 """
 Train a 7B OLMo model. Run this script without any arguments to see usage info.
 """
+
 import math
 from datetime import datetime
 
@@ -40,7 +41,10 @@ def build_model_config(common: CommonComponents) -> TransformerConfig:
     )
 
     config.block.attention.sliding_window = SlidingWindowAttentionConfig(
-        force_first=False, pattern=[False, False, False, True]
+        force_full_attention_on_first_layer=False,
+        force_full_attention_on_last_layer=True,
+        # NOTE: 4097 instead of 4096 to reproduce with the off-by-one bug.
+        pattern=[4097, 4097, 4097, -1],
     )
     config.block.attention.use_flash = True