+++ src/scripts/train/OLMo3-7B.py
@@ -57,7 +57,8 @@ def build_model_config(common: CommonComponents) -> TransformerConfig:
     config.block.attention.sliding_window = SlidingWindowAttentionConfig(
         force_full_attention_on_first_layer=False,
         force_full_attention_on_last_layer=True,
-        pattern=[4096, 4096, 4096, -1],
+        # NOTE: 4097 instead of 4096 to reproduce with the off-by-one bug.
+        pattern=[4097, 4097, 4097, -1],
     )
     config.block.attention.use_flash = True
     config.block.attention.use_head_qk_norm = True
@@ -84,6 +85,7 @@ def build_train_module_config(common: CommonComponents) -> TransformerTrainModul
             ],
             compile=False,
             foreach=True,
+            step_increment_bugfix=False,
         ),
         compile_model=True,
         dp_config=TransformerDataParallelConfig(